requirements.txt
# requirements.txt

# Interface
streamlit>=1.35.0

# Dados de mercado
yfinance>=0.2.40
requests>=2.31.0

# Cálculo numérico
numpy>=1.26.0
pandas>=2.2.0

# Visualização
plotly>=5.22.0

# Agente de IA
anthropic>=0.28.0

# Variáveis de ambiente
python-dotenv>=1.0.0
#IA
agno>=1.0.0
google-genai>=1.0.0